requirements.txt
PyYAML==3.10
distribute==0.6.34
python-dateutil==2.2
six==1.6.1
